MMSQLDB/requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.4
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
mysql-connector-python==8.1.0
mysqlclient==2.2.0
packaging==23.1
protobuf==4.21.12
python-decouple==3.8
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
typing-extensions==4.7.1
uritemplate==4.1.1
